🧑‍💻 Result컴포넌트 내용 일부를 All, List컴포넌트로 분리
